% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FFS-R90X36L9.R, R/data.R
\docType{data}
\name{Ages}
\alias{Ages}
\title{Lookup table to convert the ageing notations to integer ages}
\format{
A data frame with 128 rows and 2 variables:
\describe{
  \item{Ages}{Ages, Character string describing age(#) and growth(+) or spawning events(s)}
  \item{Int.Aget}{Integer Age, Integer converted from the Age character string}
}

A data frame with 128 rows and 2 variables:
\describe{
  \item{Ages}{Ages, Character string describing age(#) and growth(+) or spawning events(s)}
  \item{Int.Aget}{Integer Age, Integer converted from the Age character string}
}
}
\source{
PAskey personal copy with .csv stored in data-raw

PAskey personal copy with .csv stored in data-raw
}
\usage{
Ages

Ages
}
\description{
An unfortunately growing lookup table for non-conforming age notations.

An unfortunately growing lookup table for non-conforming age notations.
}
\keyword{datasets}
