% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_mat.R
\name{contrast_mat}
\alias{contrast_mat}
\title{contrast_mat}
\usage{
contrast_mat(minN = 10)
}
\arguments{
\item{minN}{the minimum sample size of fish captured in a cohort to be included in the summary.
The tally in the matrix in by lake-year, but the sample size filter is at the age-class level. So to be counted in the matrix at least one age class is above the minN in that lake-year.}
}
\description{
A function to create a pairwise comparison matrix
}
\details{
A table to show all the pairwise comparisons for the contrast of interest (tally of lake-years) in the database
}
\examples{
contrast_mat()
}
\keyword{contrast}
\keyword{matrix}
