% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkClips.R
\name{linkClips}
\alias{linkClips}
\title{SPDTdata}
\usage{
linkClips()
}
\description{
A function to load, clean and standardize SLD data for SPDT analysis. 
Function only usable by FFSBC staff who have a direct or vpn connection to SLD.
}
\details{
This is the second data filtering and cleaning process after SLD2R() to reduce down to stocking data that can be tied to releases. 
Takes SLD2R() data and uses clip information to tie fish back to stocking event, age, strain, gentotype where possible.
Make sure your VPN is running, so that the database can be accessed by the function.
The Biological record count remains the same, but includes information about the potential stocking event(s) tied to each fish.
IN cases where clips are unique, then fields for age, strain, genotype are updated if black. Otherwise a list of possibilities can be found
in "clipAges, clipStrains, clipGenos. 
All other data tables only include data that can be linked to the Biological data, either same Assessment_Key, or rel_id.
Lookup tables for integer ages and strain codes are included as part of package and can by called as Ages, Strain_code_LU
Ultimately, as upload filters and cleaning are improved in the main database, this function will become obsolete.
}
\examples{
linkClips()
}
\keyword{SPDT}
