% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FFS-R90X36L9.R, R/data.R
\docType{data}
\name{RIC_param}
\alias{RIC_param}
\title{A set of parameters to estimate relative vulnerability to capture by gillnet dependent on fish size}
\format{
A list with 4 parts: RIC_meshes, theta, rel.power are for using predict_Millar() and p_gam are stored gam model predictions
\describe{
  \item{RIC_meshes}{A vector of standard meshes in in used in RIC gillnets, numeric value}
  \item{theta}{A vecotr of 5 parameters needed for bilognorm Millar model}
  \item{rel.power}{A vector of 7 parameters to describe relative fishing power of each mesh for bilognorm Millar model}
  \item{p_gam}{A named vector describing the relative vulnerability to capture by gillnets for each fork length (mm) using a GAM model on mark-recap data} 
}

A list with 4 parts: RIC_meshes, theta, rel.power are for using predict_Millar() and p_gam are stored gam model predictions
\describe{
  \item{RIC_meshes}{A vector of standard meshes in in used in RIC gillnets, numeric value}
  \item{theta}{A vecotr of 5 parameters needed for bilognorm Millar model}
  \item{rel.power}{A vector of 7 parameters to describe relative fishing power of each mesh for bilognorm Millar model}
  \item{p_gam}{A named vector describing the relative vulnerability to capture by gillnets for each fork length (mm) using a GAM model on mark-recap data} 
}
}
\source{
PAskey personal copy. Estiamtion process described in Gillnet selectivity RMarkdown document on tech committees sharepoint.

PAskey personal copy. Estiamtion process described in Gillnet selectivity RMarkdown document on tech committees sharepoint.
}
\usage{
RIC_param

RIC_param
}
\description{
Aids in standardization of strain comparisons when groups are different sizes.

Aids in standardization of strain comparisons when groups are different sizes.
}
\keyword{datasets}
