% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby2age.R
\name{sby2age}
\alias{sby2age}
\title{sby2age}
\usage{
sby2age(Spp_code = NULL, sby_code = NULL, Year = NULL)
}
\arguments{
\item{Spp_code}{an character string or vector describing a standard species code.}

\item{sby_code}{an integer value or vecotr descibing the brood year}

\item{Year}{an integer value or vecotr descibing the year of sample (to be aged)}
}
\description{
A function to convert sby_code (brood year) to age (Jan 1st birthday convention)
}
\details{
This function calculates the age of fish based on  the species-specific spawn period and the brood year.
}
\examples{
Must be connected to VPN if working remotely

sby_code = 2005
Year = 2010
sby2age("RB", sby_code, Year)
sby2age("KO", sby_code, Year)

}
\keyword{SPDT;}
\keyword{age;}
\keyword{brood}
\keyword{year}
