% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDT-package.R
\docType{package}
\name{SPDT-package}
\alias{SPDT-package}
\title{SPDT: Stocking Product Development Team analyses (SPDT)}
\description{
This pacakge is used by SPDT members for standardized data extraction and analysis functions.
}
\keyword{internal}
