% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDT.R
\docType{package}
\name{SPDT}
\alias{SPDT}
\title{SPDT}
\description{
\code{SPDT} package to support SPDT data analyses. Source code at: \href{https://github.com/PAskey/SPDT}{SPDT website}
Explanation in \href{https://github.com/PAskey/SPDT/blob/main/README.md}{readme file}
}
\details{
Installation and Updates
Install and update from PAskey github account with devtools package. 

First install the devtools package from CRAN with:
install.packages("devtools")

Next install the SPDT package from PAskey Github repo as:
devtools::install_github("PAskey/SPDT").
}
