% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICselect.R, R/SPDTdata.R
\name{SPDTdata}
\alias{SPDTdata}
\alias{RICselect}
\title{RICselect}
\usage{
RICselect(Length_mm)

SPDTdata(Species = "RB", Contrast = "Strain", Metric = "Growth")
}
\arguments{
\item{Species}{species (use standard code in quotations) to select, default is "RB"}

\item{Contrast}{contrast of interest can be any of: "Genotype", "Strain", "SAR". Default is "Strain".}

\item{Metric}{Performance metric of interest, can be any of: "Growth" (default), "Survival" or "Maturation"}

\item{panels}{a vector of mesh sizes used in the gillnet assessment. This is assumed constant across nets.}
}
\description{
A function to correct for gillnet selectivity

A function to filter and format SLD data to cases with clipped experimental stocking groups. 
Function only usable by FFSBC staff who have a direct or vpn connection to SLD.
}
\details{
This function estimates the relative size-dependent vulnerability of trout to the standard RIC gillnets.
The function is used to multiply observed catches by a multiplier proportional to the fish fork length. 
In the future the selectivity function can be made more robust to account for non-standard combinations of gillnet panels.
Using a simply multiplier to correct for selectivity, only works with good sample sizes, because 0 * anything = 0.

This is the final data filtering and cleaning process after SLD2R(), and linkClips(). 
These two previous functions are called within this function, so there is no need to repeat.
This function returns the data tables from linkClips, plus new refined data sets.
"idf" is the individual level data frame, for all clipped fish meeting your filtering selections.
"gdf" is the grouped data frame, for all clipped groups meeting your filtering selections.
Our database will not likely ever have filtered summaries to this level, so this function will be used and updated in to the future.
#Data is filtered by the experimental contrast (Genotype, Strain, SAR) being investigated,
and the performance metric of interest (Growth, Survival, Maturation).
}
\examples{
RICselect()
SPDTdata()
}
\keyword{SPDT}
