% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDTfilter.R
\name{SPDTfilter}
\alias{SPDTfilter}
\title{SPDTfilter}
\usage{
SPDTfilter(
  Projects = NULL,
  Lk_yrs = NULL,
  Waterbody_Names = NULL,
  WBIDs = NULL,
  Years = NULL,
  Regions = NULL
)
}
\arguments{
\item{Projects}{an optional character string or character vector describing the Project_Name from Assessments table in SLD.}

\item{Lk_yrs}{an optional character string or character vector of the Lake-years to filter the data set to. Must be in format WBID_YYYY (e.g. "01100OKAN_2020" or for multiples c("01100OKAN_2020", "01598LNTH_2018"))}

\item{WBIDs}{an optional character string or character vector describing the WBIDs to include in data.}

\item{Years}{an optional integer or vector describing the sampling years to include in data.}

\item{Regions}{an optional character string or character vector describing the management regions to include in data.}

\item{Lake_Names}{an optional character string or character vector describing the lakes to include in data.}
}
\description{
A function to facilitate rapid filtering SPDTdata to specific cases beyond the biological references (lakes, years, project)
}
\details{
This is convenience data filtering and cleaning process when doing exploratory analysis or plotting with SPDTdata.
}
\examples{
#Must be connected to VPN if working remotely

#Download all data with clipped fish
SPDTdata(Contrast = "Strain")

#Creates a character vector of lake-years that exist in SPDT data call that were sampled in Mathew lake.
SPDTfilter(Waterbody_Names = "MATHEW")

#Typically we would call this within the SPDTdata (or SPDTplot) call
SPDTdata(Contrast = "Strain", filters = SPDTfilter(Waterbody_Names = "MATHEW"))

}
\keyword{SPDT}
