% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDTreleases.R
\name{SPDTreleases}
\alias{SPDTreleases}
\title{SPDTreleases}
\usage{
SPDTreleases()
}
\description{
A function to load, clean and standardize release data for SPDT analysis. 
Function only usable by staff who have a connection to SLD.
}
\details{
This is a non-exported funciton to simplify code using the releases table to understand all possible stocked groups that could be in a lake at any given time.
In cases where clips are unique, then fields for age, strain, genotype are updated.
This function was previously withing lines of exported functions: SLD2R() and linkClips().But to make code more tractable, it has been put into it's own function.
Ultimately, as upload filters and cleaning are improved in the main database, this function could become obsolete.
}
\examples{
#' Must be connected to VPN if working remotely

SPDTreleases()
}
\keyword{SPDT;}
\keyword{clips}
\keyword{releases;}
\keyword{stocked;}
