% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FFS-R90X36L9.R, R/data.R
\docType{data}
\name{Spp_code_group_LU}
\alias{Spp_code_group_LU}
\title{Lookup table with species codes and additional column for grouping species according to whether a sportfish, coarse fish, invasive, etc.}
\format{
A data frame with 170 rows and 5 variables:
\describe{
  \item{species_id}{Species identification key from SMall Lakes Database (SLD) numeric}
  \item{species_former_name}{Character string describing any former names fo rthe same species}
  \item{species_name}{charater string of current accpeted species name}
  \item{species_code}{character string for standard ( BC RISC) accepted code abbreviation for species name}
  \item{sportfish_group}{character string classifying species by the recreational, invasive, etc. values}
}

A data frame with 170 rows and 5 variables:
\describe{
  \item{species_id}{Species identification key from SMall Lakes Database (SLD) numeric}
  \item{species_former_name}{Character string describing any former names fo rthe same species}
  \item{species_name}{charater string of current accpeted species name}
  \item{species_code}{character string for standard ( BC RISC) accepted code abbreviation for species name}
  \item{sportfish_group}{character string classifying species by the recreational, invasive, etc. values}
}
}
\source{
PAskey personal copy with .csv stored in data-raw

PAskey personal copy with .csv stored in data-raw
}
\usage{
Spp_code_group_LU

Spp_code_group_LU
}
\description{
Aids in describing the recreational value or competitive nature of species composition in a lake.

Aids in describing the recreational value or competitive nature of species composition in a lake.
}
\keyword{datasets}
