% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Strain_code_LU}
\alias{Strain_code_LU}
\title{Lookup table to convert the strain names to codes}
\format{
A data frame with 19 rows and 2 variables:
\describe{
  \item{stock_strain_loc_name}{Strain name, character string fully describing strain source population}
  \item{Strain}{Strain code, Codes used to describe the strains}
}
}
\source{
PAskey personal copy
}
\usage{
Strain_code_LU
}
\description{
Aids in standardization between SLD and Paris.
}
\keyword{datasets}
