% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueClip.R
\name{uniqueClip}
\alias{uniqueClip}
\title{uniqueClip}
\usage{
uniqueClip(Clip, Species, Year, WBID, t_minus = 6)
}
\arguments{
\item{Clip}{a character string describing a specific fin clip}

\item{Species}{character string following BC standard species codes}

\item{Year}{an integer value of the assessment year to be checked}

\item{WBID}{A unique waterbody identifier for the lake being assessed}

\item{t_minus}{A number to specify how far back to check for the same clip (default is 6)}
}
\description{
A convenience function to test whether a clip is unique
}
\details{
This function is used to assess whether a fish captured with a clip is a unique group or was the clip repeated.must be a certain age class, or
It checks the historical stocking records of a lake to see if the same clip exists in a specified number of years
}
\examples{
Must be connected to VPN if working remotely, and must have already loaded the Releases data table
SD2R()
uniqueClip("AD","RB",2018, "01598LNTH")

}
\keyword{SPDT;}
\keyword{releases}
\keyword{stocked;}
