% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_to_latlong.R
\name{UTM_to_latlong}
\alias{UTM_to_latlong}
\title{UTM_to_latlong}
\usage{
UTM_to_latlong(x, y, z)
}
\arguments{
\item{x, y, z}{x = utm easting, y = utm northing, z = zone}
}
\description{
A function to convert from UTM to lat long
}
\details{
Lat Long format is prefferred for some mapping functions. This function is used within SLD2R().
}
\examples{
x = 407985
y = 5937037
z = 10
UTM_to_latlong(x,y,z)
}
\keyword{UTM}
\keyword{latlong}
