% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICselect.R
\name{RICselect}
\alias{RICselect}
\title{RICselect}
\usage{
RICselect(FLengths_mm, Millar_model = FALSE, meshSizes_in = NULL)
}
\arguments{
\item{Millar_model}{a TRUE/FALSE to indicate whether Millar model should be used for predictions}

\item{meshSizes_in}{a vecotr of mesh sizes in inches. Only applicable to Millar model and defaults to full RIC net}

\item{FLengths}{an integer or vector of fork lengths in mm for which to calculate relative probability of capture by RIC gillnet.}
}
\description{
A function to correct for gillnet selectivity
}
\details{
This function estimates the relative size-dependent vulnerability of trout to the standard RIC gillnets.
The function is used to multiply observed catches by a multiplier proportional to the fish fork length. 
In the future the selectivity function can be made more robust to account for non-standard combinations of gillnet panels.
Using a simple multiplier to correct for selectivity only works with good sample sizes, because 0 * anything = 0.
}
\examples{
Must be connected to VPN if working remotely


#Create a vector of fish lengths from 50 to 650mm
Fish_lengths = c(50:650)

#Estimate the relative probability of capture for each of the fish lengths
pvals = RICselect(Fish_lengths)

#Plot the selectivity function
plot(pvals~Fish_lengths)

#However, you can access this selectivity function directly by simply typing
select_lookup

}
\keyword{SPDT;}
\keyword{gillnet;}
\keyword{selectivity}
