% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR_cat.R
\name{SAR_cat}
\alias{SAR_cat}
\title{SAR_cat}
\usage{
SAR_cat(SAR)
}
\arguments{
\item{SAR}{a number or vector of fish release sizes (in g or other sizes)}
}
\description{
A function to partition release sizes into discrete categories
}
\details{
This function takes a number (release size) and assigns it to another number category. 
This is useful since a difference of 1g between stocking groups is a big deal for a 1g fry release size, but a rounding error for large yearlings.
A default set of categories has been created. The function may be updated to offer options in the future.
}
\examples{
Must be connected to VPN if working remotely


#Create a vector of size-at-release values - SARs
SARs = c(1:100)

#Re-categorize these into broader size bins depending on sizes. 
Default bins based on common release sizes and release size targets.
SAR_cat = SAR_cat(SARs)
SAR_cat



}
\keyword{SPDT;}
\keyword{gillnet;}
\keyword{release}
\keyword{size}
\keyword{stocking,}
