% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{select_lookup}
\alias{select_lookup}
\title{Lookup table to find relative vulnerability to capture by gillnet dependent on fish size}
\format{
A data frame with 700 rows and 2 variables:
\describe{
  \item{Length_mm}{Fork length in mm, Integer value}
  \item{p}{Relative probability of capture, based on GAM model stop gap method until more mark-recap information available}
}
}
\source{
PAskey personal copy
}
\usage{
select_lookup
}
\description{
Aids in standardization of strain comparisons when groups area  different size.
}
\keyword{datasets}
