% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLD2R.R
\name{SLD2R}
\alias{SLD2R}
\title{SLD2R}
\usage{
SLD2R()
}
\description{
A function to load, clean and standardize SLD data for SPDT analysis
}
\details{
This is the first step in doing and SPDT analyses. Loads, cleans all standard data tables for SPDT type analysis
Make sure your VPN is running, so that the database can be accessed by the function.
Lookup tables for integer ages and strain codes are included as part of package and can by called as Ages, Strain_code_LU
}
\examples{
SLD2R()
}
\keyword{SPDT}
