% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDTdata.R
\name{SPDTdata}
\alias{SPDTdata}
\title{SPDTdata}
\usage{
SPDTdata()
}
\description{
A function to filter and format SLD data to cases with clipped experimental stocking groups. 
Function only usable by FFSBC staff who have a direct or vpn connection to SLD.
}
\details{
This is the final data filtering and cleaning process after SLD2R(), and linkClips(). 
These two previous functions are called within this function, so there is no need to repeat.
This function returns the data tables from linkClips, plus new refined data sets.
"idf" is the individual level data frame, for all clipped fish meeting your filtering selections.
"gdf" is the grouped data frame, for all clipped groups meeting your filtering selections.
Our database will not likely ever have filtered summaries to this level, so this function will be used and updated in to the future.
}
\examples{
#' Must be connected to VPN if working remotely

SPDTdata()
}
\keyword{SPDT}
