% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDTdata.R
\name{SPDTdata}
\alias{SPDTdata}
\title{SPDTdata}
\usage{
SPDTdata(
  Spp = NULL,
  Contrast = NULL,
  Strains = NULL,
  Project = NULL,
  Lk_yrs = NULL,
  Data_source = c("SLD", "Biological")
)
}
\arguments{
\item{Spp}{an optional character string or character vector for BC species code (e.g. "RB" or c("KO", "EB"), etc.). This will filter data to only that species.}

\item{Contrast}{an optional character string describing the experimental contrast, which must be a field in the SPDTdata (e.g. "Strain", "SAR_cat", "Genotype" are the 3 possibilities now).
Entering a value for contrast will filter to lake years that had fish present from a co-stocking event of groups varying in your contrast variable.}

\item{Project}{an optional character describing the Project_Name from Assessments table in SLD.}

\item{Lk_yrs}{an optional character string or character vector of the Lake-years to filter the data set to. Must be in format WBID_YYYY (e.g. "01100OKAN_2020" or for multiples c("01100OKAN_2020", "01598LNTH_2018"))}

\item{Data_source}{an optional character string or file name to source the Biological data. The default "SLD" sources data from the Small Lakes Database using SPDT::linkClips(), which can take some time. 
However, using "Biological" uses Biological data in the RStudio Environment, Since the same Biological table is downloaded regardless of other options this allows user to stop reloading from the SLD with each query.
It also allows the option of the user loading a Biological Table from another source like a csv, but it must exactly match the fields in the Biological table that is produced by SPDT. 
It would be easiest to use SPDT to create a csv table and then manipulate if you wish to use data outside of the SLD.}

\item{Strain}{an optional character string or character vector describing the strain code (SPDTdata format e.g. "RB" for Rainbow Trout) for source population. This will filter to only those strains listed}
}
\description{
A function to filter and format SLD data to cases with clipped experimental stocking groups. 
Function only usable by FFSBC staff who have a direct or vpn connection to SLD.
}
\details{
This is the final data filtering and cleaning process after SLD2R(), and linkClips(). 
These two previous functions are called within this function, so there is no need to repeat.
This function returns the data tables from linkClips, plus new refined data sets.
"idf" is the individual level data frame, for all clipped fish meeting your filtering selections.
"gdf" is the grouped data frame, for all clipped groups meeting your filtering selections.
Our database will not likely ever have filtered summaries to this level, so this function will be used and updated in to the future.

Several new columns are present. Anything prefixed by "NetX" means that variable has been expanded to account for gillnet selectivity.
SAR if for size-at-release in grams, and SAR_cat has categorized those release sizes.
The code creating categories for the SAR_cat has bigger categories as the size increases and is as follows:

dplyr::mutate(SAR_cat = dplyr::case_when(SAR<=6 ~ plyr::round_any(SAR,1),
                                        SAR>6&SAR<=14.5 ~ plyr::round_any(SAR,2),
                                        SAR>14.5&SAR<=27.5 ~ plyr::round_any(SAR,5),
                                        SAR>27.5&SAR<65 ~ plyr::round_any(SAR,10),
                                        TRUE ~ plyr::round_any(SAR, 25))
You could create our own categorical system and column if needed
}
\examples{
#Must be connected to VPN if working remotely

#Download all data with clipped fish
SPDTdata()

#Download only KO data
SPDTdata(Spp = "KO")

#Download any data from RB, of either HF, CL or BW strain that were stocked as a strain comparison
SPDTdata(Spp = "RB", Strains = c("HF", "CL", "BW"), Contrast = "Strain")

#Download all data from lake years that had a size-at-release comparison for RB
SPDTdata(Spp = "RB", Contrast = "SAR_cat")

}
\keyword{SPDT}
