% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FFS-R90X36L9.R, R/data.R
\docType{data}
\name{Strain_code_LU}
\alias{Strain_code_LU}
\title{Lookup table to convert the strain names to codes}
\format{
A data frame with 235 rows and 4 variables:
<<<<<<< HEAD
\describe{
  \item{Species}{Species code, Codes used to describe the species}
  \item{stock_strain_loc_name}{Strain name, character string fully describing strain source population}
  \item{Strain}{Strain code, Codes used to describe the strains}
  \item{N}{Observations, AN integer describing the number of releases of this strain up to Jan 2022}
}

A data frame with 19 rows and 2 variables:
=======
>>>>>>> 154b74c03a9b50d1b63a556b345eb6bf1dcefc1d
\describe{
  \item{Species}{Species code, Codes used to describe the species}
  \item{stock_strain_loc_name}{Strain name, character string fully describing strain source population}
  \item{Strain}{Strain code, Codes used to describe the strains}
  \item{N}{Observations, AN integer describing the number of releases of this strain up to Jan 2022}
}
}
\source{
PAskey personal copy with .csv stored in data-raw

PAskey personal copy with .csv stored in data-raw
}
\usage{
Strain_code_LU

Strain_code_LU
}
\description{
Aids in standardization between SLD and Paris.

Aids in standardization between SLD and Paris.
}
\keyword{datasets}
